requirements.txt
annotated-types==0.6.0
asgiref==3.6.0
beautifulsoup4==4.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
Django==4.2
django-bootstrap-datepicker-plus==5.0.5
fonttools==4.42.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.7.2
nibabel==5.1.0
numpy==1.25.2
opencv-python==4.8.1.78
packaging==23.1
Pillow==10.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.15
